updated the install script

+++ install/install.sh
@@ -1,3 +1,7 @@
+if ! pacman -Q git >> /dev/null; then
+  sudo pacman -S --noconfirm git
+fi
+
 #!/usr/bin/env bash
 sudo -v
 
@@ -10,6 +14,9 @@ fi
 
 if pacman -Q pulseaudio >> /dev/null; then
   sudo pacman -Rc --noconfirm pulseaudio pulseaudio-alsa pulseaudio-bluetooth
+fi
+
+if ! pacman -Q pipewire >> /dev/null; then
   sudo pacman -S --noconfirm pipewire pipewire-pulse pipewire-audio
 fi
 
@@ -42,85 +49,95 @@ else
   fi
 fi
 
-sudo pacman-key --recv-key 3056513887B78AEB --keyserver keyserver.ubuntu.com
-sudo pacman-key --lsign-key 3056513887B78AEB
-sudo pacman -U --noconfirm 'https://cdn-mirror.chaotic.cx/chaotic-aur/chaotic-keyring.pkg.tar.zst'
-sudo pacman -U --noconfirm 'https://cdn-mirror.chaotic.cx/chaotic-aur/chaotic-mirrorlist.pkg.tar.zst'
+sudo -v
 
-sudo bash -c "echo -e \"\n[chaotic-aur]\" >> \"/etc/pacman.conf\""
-sudo bash -c "echo -e \"Include = /etc/pacman.d/chaotic-mirrorlist\" >> \"/etc/pacman.conf\""
+if ! pacman -Q chaotic-keyring >> /dev/null && pacman -Q chaotic-keyring >> /dev/null; then
+  sudo pacman-key --recv-key 3056513887B78AEB --keyserver keyserver.ubuntu.com
+  sudo pacman-key --lsign-key 3056513887B78AEB
+  sudo pacman -U --noconfirm 'https://cdn-mirror.chaotic.cx/chaotic-aur/chaotic-keyring.pkg.tar.zst'
+  sudo pacman -U --noconfirm 'https://cdn-mirror.chaotic.cx/chaotic-aur/chaotic-mirrorlist.pkg.tar.zst'
 
-if pacman -Q hyprland >> /dev/null; then
-  echo "Normal packages are installed."
-else
-  sudo pacman -Syyu --noconfirm sway \
-                                htop \
-                                papirus-icon-theme \
-                                waybar \
-                                eza \
-                                qt5-wayland \
-                                swaybg \
-                                blueberry \
-                                swaync \
-                                wdisplays \
-                                wl-clipboard \
-                                grim \
-                                slurp \
-                                bat-extras \
-                                fish \
-                                starship \
-                                pcmanfm-gtk3 \
-                                file-roller \
-                                p7zip \
-                                unrar \
-                                zathura \
-                                zathura-pdf-mupdf \
-                                gammastep \
-                                lolcat \
-                                ttf-ubuntu-nerd \
-                                ttf-jetbrains-mono-nerd \
-                                noto-fonts-emoji \
-                                pavucontrol \
-                                qt5ct \
-                                foot \
-                                networkmanager \
-                                qutebrowser \
-                                python-adblock \
-                                mpv \
-                                mpv-mpris \
-                                polkit-gnome \
-                                grimshot \
-                                dracula-cursors-git \
-                                light \
-                                chafa \
-                                ripgrep \
-                                fzf \
-                                hyprpicker-git \
-                                swayimg \
-                                emacs-wayland \
-                                lf \
-                                stow \
-                                expac \
-                                python-tldextract \
-                                python-pynacl \
-                                fastfetch \
-                                git \
-                                tree-sitter \
-                                udiskie \
-                                swaylock
-                                # hyprland \
-                                # neovim \
-                                # otf-codenewroman-nerd \
+  if ! grep "chaotic" /etc/pacman.conf >> /dev/null; then
+    echo -e "\n[chaotic-aur]" | sudo tee -a /etc/pacman.conf
+    echo -e "Include = /etc/pacman.d/chaotic-mirrorlist" | sudo tee -a /etc/pacman.conf
+  fi
 fi
 
 if pacman -Q yay >> /dev/null; then
   echo "It seems yay is installed."
 else
+  sudo pacman -S --noconfirm base-devel
   git clone https://aur.archlinux.org/yay-bin
   cd yay-bin
   makepkg -si --noconfirm
   cd ..
   rm -rf yay-bin/
+fi
+
+sudo -v
+
+if pacman -Q waybar >> /dev/null; then
+  echo "Normal packages are installed."
+else
+  yay -Syyu --noconfirm sway \
+                        htop \
+                        papirus-icon-theme \
+                        waybar \
+                        eza \
+                        qt5-wayland \
+                        swaybg \
+                        blueberry \
+                        swaync \
+                        wdisplays \
+                        wl-clipboard \
+                        grim \
+                        slurp \
+                        bat-extras \
+                        fish \
+                        starship \
+                        pcmanfm-gtk3 \
+                        file-roller \
+                        p7zip \
+                        unrar \
+                        zathura \
+                        zathura-pdf-mupdf \
+                        gammastep \
+                        lolcat \
+                        ttf-ubuntu-nerd \
+                        ttf-jetbrains-mono-nerd \
+                        noto-fonts-emoji \
+                        pavucontrol \
+                        qt5ct \
+                        foot \
+                        networkmanager \
+                        python-adblock \
+                        mpv \
+                        mpv-mpris \
+                        polkit-gnome \
+                        grimshot \
+                        dracula-cursors-git \
+                        light \
+                        chafa \
+                        ripgrep \
+                        fzf \
+                        swayimg \
+                        emacs-wayland \
+                        lf \
+                        stow \
+                        expac \
+                        python-tldextract \
+                        python-pynacl \
+                        fastfetch \
+                        tree-sitter \
+                        udiskie \
+                        xorg-xwayland \
+                        brave-bin
+                        # qutebrowser \
+                        # swaylock
+                        # hyprpicker-git \
+                        # hyprland \
+                        # neovim \
+                        # otf-codenewroman-nerd \
 fi
 
 if pacman -Q rofi-lbonn-wayland-git >> /dev/null; then
@@ -143,6 +160,8 @@ else
                      # bitwarden-rofi-git \
 fi
 
+sudo -v
+
 if pacman -Q wpgtk >> /dev/null; then
   echo "It seems pywal packages are installed."
 else
@@ -153,7 +172,7 @@ else
                      python-prctl \
                      python-daemon \
                      python-haishoku \
-                     chaotic-aur/qt5-styleplugins \
+                     qt5-styleplugins \
                      chaotic-aur/qt6gtk2 \
                      python-zombie-imp \
                      gradience \
@@ -162,17 +181,19 @@ else
                      # chaotic-aur/adw-gtk3
 
   # using pip like this can break stuff, be careful
-  sudo pip install yapsy --break-system-packages
-  sudo pip install anyascii --break-system-packages
+  # sudo pip install yapsy --break-system-packages
+  # sudo pip install anyascii --break-system-packages
   sudo pip install modern_colorthief --break-system-packages
 
   wpg-install.sh -g
 fi
 
+sudo -v
+
 if pacman -Q picard >> /dev/null; then
   echo "Optional packages are installed."
 else
-  yay -S --noconfirm appimagelauncher-bin \
+  yay -S --noconfirm appimagelauncher \
                      keepassxc \
                      syncthing \
                      gnome-disk-utility \
@@ -186,6 +207,8 @@ else
                      ntfs-3g
                      # syncthing-gtk
 fi
+
+sudo -v
 
 if [ -e "/etc/X11/xorg.conf.d/90-touchpad.conf" ]; then
   echo "It seems you already have touchpad configuration. No changes have been made."
@@ -206,6 +229,8 @@ EndSection'
   echo "The changes have been applied. After reboot touchpad should work."
 fi
 
+sudo -v
+
 if pacman -Q texlive-bin >> /dev/null; then
   echo "Latex is setup."
 else
@@ -214,15 +239,19 @@ else
   texhash
 fi
 
-git clone https://gitlab.com/imnotpua/grub_gtg
+sudo -v
 
-cd grub_gtg
+if [ ! -e "/boot/grub/themes/grub_gtg" ]; then
+  git clone https://gitlab.com/imnotpua/grub_gtg
 
-echo "YOU NEED TO TYPE FONT SIZE AND THEME DIRECTORY"
-sudo bash ./install.sh
+  cd grub_gtg
 
-cd ../
-rm -rf grub_gtg
+  echo "YOU NEED TO TYPE FONT SIZE AND THEME DIRECTORY"
+  sudo bash ./install.sh
+
+  cd ../
+  rm -rf grub_gtg
+fi
 
 chsh -s "$(command -v fish)"
 curl https://raw.githubusercontent.com/oh-my-fish/oh-my-fish/master/bin/install -o "omf"
@@ -230,7 +259,9 @@ fish omf --noninteractive
 fish -c "omf install foreign-env"
 rm omf
 
-xdg-settings set default-web-browser org.qutebrowser.qutebrowser.desktop
+if pacman -Q qutebrowser >> /dev/null; then
+  xdg-settings set default-web-browser org.qutebrowser.qutebrowser.desktop
+fi
 
 gsettings set org.gnome.desktop.interface gtk-theme "FlatColor"
 gsettings set org.gnome.desktop.interface icon-theme "Papirus-Dark"
@@ -244,4 +275,6 @@ yay -S --noconfirm xdg-user-dirs
 xdg-user-dirs-update
 
 # enabling syncthing service
-systemctl --user enable syncthing
+if pacman -Q syncthing >> /dev/null; then
+  systemctl --user enable syncthing
+fi
